+++ ios/Classes/AdPopcornRewardPlugin.h
@@ -1,5 +1,6 @@
 #import <Flutter/Flutter.h>
 #import <AdPopcornOfferwall/AdPopcornOfferwallSDK.h>
+#import "AdPopcornRewardFLNativeView.h"
 
 @interface AdPopcornRewardPlugin : NSObject<FlutterPlugin>
 @property(nonatomic, strong) FlutterMethodChannel *channel;